Add report command
The report command lists all logged entries of the current day, the sums
of all logged types and a total sum.
Using the reportDay argument a specific day for the report command can
be set, e.g.:

    clocker report --reportDay 2018-05-04

If no report day was set, the report will be displayed for the current day.